requirements.txt
AWSIoTPythonSDK
pyserial
pynmea2
